% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{accel_summaries}
\alias{accel_summaries}
\title{Summarize accelerometer files by device}
\usage{
accel_summaries(
  device,
  data_folder,
  output_folder,
  epochs = c(60),
  sample_rate = 100,
  dynamic_range = c(-8, 8),
  apply_nonwear = FALSE,
  nonwear_cfg = list(min_bout = 90L, spike_tol = 2L, spike_upper = 100L, flank = 30L),
  tz = "UTC"
)
}
\description{
Summarize accelerometer files by device
}
